combine scheduler and nightclock
